// src/shared/services/api/models/courseContentResponseDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Learning Assessment Portal API
 * OpenAPI spec version: v1
 */
import type { CourseTopicProgressDto } from './courseTopicProgressDto';

export interface CourseContentResponseDto {
  course_id?: string;
  /** @nullable */
  thumbnail_img?: string | null;
  /** @nullable */
  topic?: CourseTopicProgressDto[] | null;
}