// src/shared/services/api/models/courseTopicProgressDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Learning Assessment Portal API
 * OpenAPI spec version: v1
 */
import type { CourseContentProgressDto } from './courseContentProgressDto';

export interface CourseTopicProgressDto {
  id?: string;
  /** @nullable */
  name?: string | null;
  meta_sequence_order?: number;
  meta_duration_minute?: number;
  /** @nullable */
  contents?: CourseContentProgressDto[] | null;
}